아키텍처 대 ㅡ 변 ㅡ 신
