+++ src/main/java/com/hhplus/concert/auth/presentation/dto/response/AuthTokenPostResVo.java
@@ -1,0 +1,14 @@
+package com.hhplus.concert.auth.presentation.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class AuthTokenPostResVo {
+    private String uuid;
+}
