--- src/main/java/com/hhplus/concert/domain/Schedule.java
@@ -1,11 +1,0 @@
-package com.hhplus.concert.domain;
-
-import java.time.LocalDateTime;
-
-public record Schedule(
-    Long scheduleId,
-    Long concertId,
-    LocalDateTime scheduleDate
-) {
-
-}
