--- src/main/java/com/hhplus/concert/domain/Seat.java
@@ -1,10 +1,0 @@
-package com.hhplus.concert.domain;
-
-public record Seat(
-    Long seatId,
-    Long seatNumber,
-    Long scheduleId,
-    Long price
-) {
-
-}
