--- src/main/java/com/hhplus/concert/infrastructure/type/SeatStatus.java
@@ -1,5 +1,0 @@
-package com.hhplus.concert.infrastructure.type;
-
-public enum SeatStatus {
-
-}
