--- src/main/java/com/hhplus/concert/infrastructure/type/TokenStatus.java
@@ -1,16 +1,0 @@
-package com.hhplus.concert.infrastructure.type;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum TokenStatus {
-    //reserved, available, expired
-    RESERVED("1", "대기"),
-    AVAILABLE("2", "입장"),
-    EXPIRED("3", "만료");
-
-    private String code;
-    private String desc;
-}
