+++ src/main/java/com/hhplus/concert/payment/presentation/dto/response/PaymentResVo.java
@@ -1,0 +1,16 @@
+package com.hhplus.concert.payment.presentation.dto.response;
+
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class PaymentResVo {
+    private Long paymentId;
+    private LocalDateTime paymentTime;
+}
