+++ src/main/java/com/hhplus/concert/payment/presentation/dto/response/UserBalanceResVo.java
@@ -1,0 +1,15 @@
+package com.hhplus.concert.payment.presentation.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class UserBalanceResVo {
+    private Long userId;
+    private Long balance;
+}
