--- src/main/java/com/hhplus/concert/presentation/dto/request/AuthTokenPostReqVo.java
@@ -1,14 +1,0 @@
-package com.hhplus.concert.presentation.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-public class AuthTokenPostReqVo {
-    private Long userId;
-}
