--- src/main/java/com/hhplus/concert/presentation/dto/response/AuthTokenPostResVo.java
@@ -1,14 +1,0 @@
-package com.hhplus.concert.presentation.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-public class AuthTokenPostResVo {
-    private String uuid;
-}
