--- src/main/java/com/hhplus/concert/presentation/dto/response/PaymentResVo.java
@@ -1,16 +1,0 @@
-package com.hhplus.concert.presentation.dto.response;
-
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-public class PaymentResVo {
-    private Long paymentId;
-    private LocalDateTime paymentTime;
-}
