--- src/main/java/com/hhplus/concert/presentation/dto/response/ReservationSeatResVo.java
@@ -1,19 +1,0 @@
-package com.hhplus.concert.presentation.dto.response;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-public class ReservationSeatResVo {
-    private Long reservationId;
-    private List<Long> seatIdList;
-    private Long totalPrice;
-    private LocalDateTime reservationExpiry;
-}
