+++ src/main/java/com/hhplus/concert/reservation/domain/Schedule.java
@@ -1,0 +1,11 @@
+package com.hhplus.concert.reservation.domain;
+
+import java.time.LocalDateTime;
+
+public record Schedule(
+    Long scheduleId,
+    Long concertId,
+    LocalDateTime scheduleDate
+) {
+
+}
