+++ src/main/java/com/hhplus/concert/reservation/domain/Seat.java
@@ -1,0 +1,10 @@
+package com.hhplus.concert.reservation.domain;
+
+public record Seat(
+    Long seatId,
+    Long seatNumber,
+    Long scheduleId,
+    Long price
+) {
+
+}
