+++ src/main/java/com/hhplus/concert/reservation/domain/type/SeatStatus.java
@@ -1,0 +1,5 @@
+package com.hhplus.concert.reservation.domain.type;
+
+public enum SeatStatus {
+
+}
