+++ src/main/java/com/hhplus/concert/reservation/presentation/dto/request/SeatReqVo.java
@@ -1,0 +1,18 @@
+package com.hhplus.concert.reservation.presentation.dto.request;
+
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class SeatReqVo {
+    private Long userId;
+    private Long contentId;
+    private Long scheduleId;
+    private List<Long> seatIdList;
+}
